+++ Sources/XCCovLib/Utils.swift
@@ -5,7 +5,8 @@
 //
 // SPDX-License-Identifier: MIT
 //
-// Originally created by David Whetstone, Trax, 10/16/19.
+// Originally created by David Whetstone @ Trax Retail, 10/16/19.
+//
 
 import Foundation
